Former-commit-id: 3626c57c40758079c905c6b2baf3ca083bc320b5 [formerly b6af57b411b9ec4c9fa3ba83714e67adf8c15db1] [formerly 29677326c7a0b9dcf115b06d8cb2e0645a2c7614] [formerly 29677326c7a0b9dcf115b06d8cb2e0645a2c7614 [formerly be88c9494ecedbdb140a4995c6feec6b29c4cfee]] [formerly 6a1714a7b00a5510b395a91ab3d4e4170651907a [formerly 29677326c7a0b9dcf115b06d8cb2e0645a2c7614 [formerly be88c9494ecedbdb140a4995c6feec6b29c4cfee] [formerly 6a1714a7b00a5510b395a91ab3d4e4170651907a [formerly f65e8a7253c65b5d59e78ed57e3d73bf01b2e06f]]]]
Former-commit-id: 6a1714a7b00a5510b395a91ab3d4e4170651907a
Former-commit-id: 76ce22b10999b668d133b7331eae90c7f6254db1 [formerly 119ca9e825f2d313a0116934bf8546a845a3fdd3] [formerly e280c293be772e3b76431146949849c8b59106f3 [formerly 44ecd4408b0a495b887d8514b0d31c9c7be47f74]]
Former-commit-id: e280c293be772e3b76431146949849c8b59106f3
Former-commit-id: a11d23c5050a27e81cfdaca00c335e8e5c1cd72c

+++ edexOsgi/com.raytheon.uf.edex.plugin.bufrascat/src/com/raytheon/uf/edex/plugin/bufrascat/decoder/AScatDataAdapter.java
@@ -50,6 +50,7 @@ import com.raytheon.uf.edex.pointdata.PointDataPluginDao;
  *                                     types.
  * Aug 30, 2013 2298       rjpeter     Make getPluginName abstract
  * May 14, 2014 2536       bclement    moved WMO Header to common, removed TimeTools usage
+ * Jul 23, 2014 3410       bclement    location changed to floats
  * 
  * </pre>
  * 
@@ -137,7 +138,7 @@ public class AScatDataAdapter extends BUFRPointDataAdapter<AScatObs> {
                             IDecoderConstants.VAL_MISSING);
 
                     SurfaceObsLocation location = new SurfaceObsLocation();
-                    location.assignLocation(lat, lon);
+                    location.assignLocation((float) lat, (float) lon);
                     location.generateCoordinateStationId();
                     obsData.setLocation(location);
